Add tests for request and response objects + change float into double for local produce count

+++ src/main/java/controller/FoodController.java
@@ -24,7 +24,7 @@ public class FoodController {
     public void updateViewWithData() {
         int vegmeals = model.getVegMealsCount();
         view.updateVegCounter(vegmeals);
-        float local = model.getLocalProduceCount();
+        double local = model.getLocalProduceCount();
         view.updateLocalProduceCounter(local);
     }
 
@@ -54,10 +54,10 @@ public class FoodController {
      * @param weightString - weight of the eaten produces in kg.
      */
     public void addEatenLocalProduce(String weightString) {
-        float kg = 0;
+        double kg = 0;
 
         try {
-            kg = Float.parseFloat(weightString);
+            kg = Double.parseDouble(weightString);
             if (kg < 0 || kg > 10) {
                 view.displayStatus("Please enter the weight in the range 0-10kg.");
                 return;
